[PlayerMap] Add buildings initi

+++ TerreHostile/TH_Business/src/main/java/org/terrehostile/repository/UserRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.terrehostile.business.admin.authentification.User;
 
-
 @Repository("userRepository")
-public interface UserRepository extends JpaRepository<User, Long> {
-	 User findByEmail(String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
+	User findByEmail(String email);
 }
